% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_categ.R
\name{add_coherence}
\alias{add_coherence}
\title{Add a column with 'coherence' information to a data.frame}
\usage{
add_coherence(data, col_name = "coherent")
}
\arguments{
\item{data}{a data.frame. Some columns are mandatory. See details for more information}

\item{col_name}{name of the new column created}
}
\value{
'data' data.frame with a new logical column name after 'col_name'
}
\description{
Internal function that calls is_iti_coherent
}
\details{
data should have the following columns: iti_km_voyage, iti_experience,
iti_depart_itineraire, iti_arrivee_itineraire,
iti_depart_initial, iti_arrivee_final
}
