% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_list_df}
\alias{bind_list_df}
\title{Column binds list of data.frames element by element}
\usage{
bind_list_df(...)
}
\arguments{
\item{...}{several named lists of data.frames}
}
\value{
a list of data.frames
}
\description{
This function will look at each element of the list y name (not by position) and will cbind the one with the same names
It returns one list with names of all the input lists
}
\keyword{internal}
