% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_evavelo.R
\name{check_diff}
\alias{check_diff}
\title{Check correctness of id_site_enq and date_enq between 2 data.frames}
\usage{
check_diff(a, b, name_a = NULL, name_b = NULL)
}
\arguments{
\item{a}{a data.frame containing id_site_enq and date_enq variables}

\item{b}{a data.frame containing id_site_enq and date_enq variables}

\item{name_a}{name of the df to be used in the log message. If missing, we will use the name of data.frame a}

\item{name_b}{name of the df to be used in the log message. If missing, we will use the name of data.frame b}
}
\value{
a log message with the differences between data.frames
}
\description{
Check correctness of id_site_enq and date_enq between 2 data.frames
}
