% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_evavelo.R
\name{check_evavelo}
\alias{check_evavelo}
\title{Check input file correctness (Eva-velo)}
\usage{
check_evavelo(calendrier, comptage, enquete, comptage_init, enquete_init)
}
\arguments{
\item{calendrier}{a data.frame corresponding to sheet "calendrier_sites"}

\item{comptage}{a data.frame corresponding to sheet "comptages_man_post_traitements"}

\item{enquete}{a data.frame corresponding to sheet "enquetes_post_traitement"}

\item{comptage_init}{a data.frame corresponding to sheet "comptages_manuels"}

\item{enquete_init}{a data.frame corresponding to sheet "enquetes_saisies"}
}
\value{
a list of two values: error (a boolean) and log (a string)
}
\description{
Check input file correctness (Eva-velo)
}
