% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_num_outliers.R
\name{check_num_outliers}
\alias{check_num_outliers}
\title{Search for outliers in numeric values}
\usage{
check_num_outliers(evadata, categorie_corrige = NULL)
}
\arguments{
\item{evadata}{an evadata object obtained with read_eva_velo()}

\item{categorie_corrige}{optional vector of categorie_corrige. Must be the same size as the number of rows in evadata$enquete.
To be used if we want to use different categorie_corrige than the one stored in evadata}
}
\value{
a list of data.frames with different outliers detected
}
\description{
Search for outliers in numeric values
}
