% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_evavelo.R
\name{check_similar_enquete}
\alias{check_similar_enquete}
\title{Detects similar response in enquete}
\usage{
check_similar_enquete(enquete)
}
\arguments{
\item{enquete}{enquete data.frame obtained with `read_enquete()`}
}
\value{
a data.frame with one columns enq_list containing potential similar answers
}
\description{
Try to detect multiple id_quest that were not tagged as
Will group identical answers to a series of text answers
And then will confirm with the similarity of numeric responses (km_sortie, heure_enq)
}
