% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_cities.R
\name{check_warn_cities}
\alias{check_warn_cities}
\title{Check a list of cities and try to find replacement values}
\usage{
check_warn_cities(data)
}
\arguments{
\item{data}{a data.frame with the following columns:
city (string), postcode(string), result_type, result_score, result_label, result_postcode.
This data.frame is produced using banR::geocode_tbl()}
}
\value{
invisible(0)
}
\description{
Information is printed in warnings. This function doesn't return any values.
}
