% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_cities_cp.R
\name{check_warn_cities_cp}
\alias{check_warn_cities_cp}
\title{Check a list of cities and try to find replacement values}
\usage{
check_warn_cities_cp(data)
}
\arguments{
\item{data}{a data.frame with the following columns:
id_rows, city (string), postcode(string), result_type, result_score, result_label, result_postcode.
This data.frame is produced using banR::geocode_tbl()}
}
\value{
invisible(0)
}
\description{
Information is printed in warnings.
This function returns a data.frame containing correction proposition
}
\keyword{internal}
