% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_init_post}
\alias{compare_init_post}
\title{Compare two data.frames (initial and  post treatment)}
\usage{
compare_init_post(init, post, key = NULL)
}
\arguments{
\item{init}{a data.frame (initial data.frame)}

\item{post}{a data.frame (post treatment data.frame)}

\item{key}{by default (NULL) the row order of both data.frame has to be the same.
by using a key column name, both data.frames are ordered before being compared.}
}
\value{
invisible(0)
}
\description{
In this comparison, all elements from init data.frame should appear inside post data.frame
post data.frame can only have new columns but no new rows.
}
\details{
This function doesn't return anything but creates warnings if there are mismatches.
}
\keyword{internal}
