% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_categ.R
\name{correct_categ}
\alias{correct_categ}
\title{Analyze and correct categorie}
\usage{
correct_categ(comptage, enquete)
}
\arguments{
\item{comptage}{a data.frame that must contain the following columns: .....}

\item{enquete}{a data.frame that must contain the following columns: .....}
}
\value{
a vector the same length as the number of rows in comptage with the new `categorie` value
}
\description{
This function takes `comptage` and `enquete` data to compute categorie_corrigee field.
}
\keyword{internal}
