% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_categ.R
\name{correct_itinerant}
\alias{correct_itinerant}
\title{Apply categorie_corrige methodology when Itinerant is present in the answer}
\usage{
correct_itinerant(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
a data.frame the same size of data with updated categorie_corrige values.
}
\description{
In Chapter 3.1.11 , this corresponds to cases 1 2 3 4 7 10
}
\details{
this function can be used inside pipe operator and is compatible with dplyr
}
\keyword{internal}
