% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_has_cols}
\alias{df_has_cols}
\title{Check that columns are present in a data.frame}
\usage{
df_has_cols(x, cols)
}
\arguments{
\item{x}{a data.frame}

\item{cols}{a columns names vector}
}
\description{
THrows an error if some columns `cols` are not present in the data.frame `x`
}
