% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{equal_with_na}
\alias{equal_with_na}
\title{Test equality between vectors including NA values}
\usage{
equal_with_na(a, b)
}
\arguments{
\item{a}{a vector}

\item{b}{a vector}
}
\value{
a boolean vector with element wise comparison
}
\description{
This function is a variant of == that takes in account NAs.
By default NA == NA returns FALSE in R
With this fonction, NAs can be compared
}
