% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{evavelo_example}
\alias{evavelo_example}
\title{evavelo example object}
\format{
An object of class \code{evadata} (inherits from \code{list}) of length 7.
}
\usage{
evavelo_example
}
\description{
A simplified evavelo object based on a subset of a real investigation
}
\keyword{datasets}
