% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{evavelo_example_geocoded}
\alias{evavelo_example_geocoded}
\title{evavelo example geocoded object}
\format{
An object of class \code{evadata} (inherits from \code{list}) of length 7.
}
\usage{
evavelo_example_geocoded
}
\description{
a geocoded version of evavelo_example
}
\keyword{datasets}
