% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_evavelo.R
\name{find_wrong_cities}
\alias{find_wrong_cities}
\title{Find cities that couldn't be geocoded by geocode_cities()}
\usage{
find_wrong_cities(var_name, enquete, fields_to_keep = "id_quest")
}
\arguments{
\item{var_name}{string corresponding to the city name variable to check}

\item{enquete}{data.frame containing the var_name variable and an associated *_cog}

\item{fields_to_keep}{fields that will be exported along with var_name}
}
\value{
a data.frame with cities that couldn't be geocoded
}
\description{
Find cities that couldn't be geocoded by geocode_cities()
}
\keyword{internal}
