% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_evavelo.R
\name{find_wrong_geocoding}
\alias{find_wrong_geocoding}
\title{Find all cities that failed to geocode when using geocode_evavelo()}
\usage{
find_wrong_geocoding(data)
}
\arguments{
\item{data}{a geocoded eva_data object}
}
\value{
a list of data.frames with all the cities that couldn't be geocoded
}
\description{
Find all cities that failed to geocode when using geocode_evavelo()
}
