% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_cities.R
\name{geocode_cities}
\alias{geocode_cities}
\title{geocode a tibble with french cities}
\usage{
geocode_cities(.data, city_col, cp_col = NULL, country_col = NULL)
}
\arguments{
\item{.data}{a dataframe that needs to be updated}

\item{city_col}{column name containing the city name in .data
This value will also be used to name the new columns.}

\item{cp_col}{optional postcode columns}

\item{country_col}{optional contry_col (to avoid geocoding non-french cities)}
}
\value{
a the input data.frame with 3 new columns with a name based on city_col (_lat, _long, _cog)
}
\description{
geocode a tibble with french cities
}
