% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_utils.R
\name{geocode_df_cities}
\alias{geocode_df_cities}
\title{geocode df for cities}
\usage{
geocode_df_cities(.data, city_col)
}
\arguments{
\item{.data}{a data.frame}

\item{city_col}{name of the column that contains the city names}
}
\value{
.data with additional columns named :
result_name, result_cog, result_lat, result_lon, result_score
}
\description{
Similar to banR::geocode_tbl() specifically for cities but with improvement:
}
\details{
Size of the request is reduced by avoiding asking several times the same city.
NA returns NA ansd not cities starting with "na"
Messages are not displayed when there is no error.
input.data is not altered even if contains special characters. Column order is also preserved
}
\keyword{internal}
