% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_cities_cp.R
\name{geocode_df_foreign_cities}
\alias{geocode_df_foreign_cities}
\title{geocode df for foreign cities}
\usage{
geocode_df_foreign_cities(.data, city_col, country_col)
}
\arguments{
\item{.data}{a data.frame}

\item{city_col}{name of the column that contains the city names}

\item{country_col}{name of the column that contains the country names}
}
\value{
.data with additional (or updated) columns named after city_col with the extension:
_lat _lon and _cog
}
\description{
A variation of geocode_df_cities for foreign cities.
Instead of using banR, we use tidygeocoder with osm (nomatim) API
}
\details{
Messages are not displayed when there is no error.
input.data is not altered even if contains special characters.
Column order is also preserved.
Countries that not "France" or NA will not be altered by this function
}
\keyword{internal}
