% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_evavelo.R
\name{geocode_evavelo}
\alias{geocode_evavelo}
\title{Geocode a eva_data object}
\usage{
geocode_evavelo(data)
}
\arguments{
\item{data}{an evadata object obtained with read_evavelo() or an xlsx file, Workbook object or URL to xlsx file}
}
\value{
a geocoded eva_data object
}
\description{
Geocode a eva_data object
}
