% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_utils.R
\name{geocode_nomatim}
\alias{geocode_nomatim}
\title{Geocode foreeign cities using Nomatim (OSM) API}
\usage{
geocode_nomatim(df, city = city, country = country)
}
\arguments{
\item{df}{a data.frame containing at least id_rows variable and 2 other with city and country information}

\item{city}{name of the column containing the city}

\item{country}{name of the column containing the city}
}
\value{
a data.frame containing 3 columns: id_rows, lat and lon
}
\description{
Wrapper around tidygeocoder::geocode() that removes unnecessary messages and progress bars
and format the output data.table with needed variables
}
