% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_utils.R
\name{geocode_row_by_row}
\alias{geocode_row_by_row}
\title{Geocode a data.frame of cities row by row}
\usage{
geocode_row_by_row(.data, city_col, score_limit = 0.4)
}
\arguments{
\item{.data}{a data.frame containing a columns with city names to geocode}

\item{city_col}{column that contains city names}

\item{score_limit}{results below this score will be ignored and replaced by NAs}
}
\value{
the input data.frame with geocoding infomation.
}
\description{
Equivalent to banR::geocode_tbl() but much slower (one API request per row)
This function is used when we only want "municipality" as result type which is impossible with geocode_tbl().
}
\keyword{internal}
