% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_table_communes.R
\name{geocode_table_communes}
\alias{geocode_table_communes}
\title{geocode table_communes cities}
\usage{
geocode_table_communes(table_communes)
}
\arguments{
\item{table_communes}{a data.frame containing columns nom_commune and COG}
}
\value{
a data.frame similar to table_communes with latitude and longitude added.
}
\description{
THis function check that the municipalities defined in tables_communes are correct
and geocodes when possible
}
