% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distance.R
\name{get_closest_point}
\alias{get_closest_point}
\title{Get closest point b from points a}
\usage{
get_closest_point(lon_a, lat_a, lon_b, lat_b, id_b, max_dist = 30, ...)
}
\arguments{
\item{lon_a}{points a longitude}

\item{lat_a}{points a latitude}

\item{lon_b}{points b longitude}

\item{lat_b}{points b latitude}

\item{id_b}{b ids used in the output. Should have the same size as lon_b and lat_b}

\item{max_dist}{distances greater than this value in kilometers will be discarded and result will have NAs}

\item{...}{additionae arguments passed to geodist::geodist_vec()}
}
\value{
a vector the same size as lon_a and lat_a
}
\description{
Output a vector with the id of points b that is the closest to points a
}
