% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distance.R
\name{intersection_traj_reg}
\alias{intersection_traj_reg}
\title{Find the entrance point inside a defined region}
\usage{
intersection_traj_reg(
  .enquete,
  region_shape,
  arrival = "nom_site_enq",
  departure = "ville_res",
  output_prefix = "result"
)
}
\arguments{
\item{.enquete}{enquete part of a geocoded eva_data object}

\item{region_shape}{sf object. A Polygon of the region to study}

\item{arrival}{names of the arrival point. the corresponding variable _lon _lat must exist in enquete
We assume that all arrival points are in the region}

\item{departure}{names of the departure point. the corresponding variable _lon _lat must exist in enquete}

\item{output_prefix}{prefix used for the output columns}
}
\value{
a data.frame the same length as enquete with 3 variables id_quest result_lon result_lat
}
\description{
If both arrival and departure are inside region, we return NA
}
\keyword{internal}
