% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_empty_df}
\alias{is_empty_df}
\title{Check if  data.frame has 0 row}
\usage{
is_empty_df(df)
}
\arguments{
\item{df}{}
}
\value{
a boolean
}
\description{
Check if  data.frame has 0 row
}
\keyword{internal}
