% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_categ.R
\name{is_iti_coherent}
\alias{is_iti_coherent}
\title{Helper function to control "coherence" of specific "itinerance" answers:}
\usage{
is_iti_coherent(
  iti_km_voyage,
  iti_experience,
  iti_dep_iti_valide,
  iti_arr_iti_valide,
  iti_depart_initial,
  iti_arrivee_final
)
}
\arguments{
\item{iti_km_voyage}{numeric vector}

\item{iti_experience}{character vector}

\item{iti_dep_iti_valide}{character vector}

\item{iti_arr_iti_valide}{character vector}

\item{iti_depart_initial}{character vector}

\item{iti_arrivee_final}{character vector}
}
\value{
a boolean vector indicating if answer is coherent
}
\description{
https://github.com/JMPivette/evavelo/discussions/7
 https://github.com/JMPivette/evavelo/issues/39
}
