% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_evavelo.R
\name{log_in_x_not_in_y}
\alias{log_in_x_not_in_y}
\title{Utils function to log date site combination that are in x and not in y}
\usage{
log_in_x_not_in_y(x, y, name_x, name_y)
}
\arguments{
\item{x}{a data.frame containing id_site_enq and date_enq variables}

\item{y}{a data.frame containing id_site_enq and date_enq variables}

\item{name_x}{name of the df to be used in the log message.}

\item{name_y}{name of the df to be used in the log message.}
}
\value{
a string which logs the differences
}
\description{
Utils function to log date site combination that are in x and not in y
}
