% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distance.R
\name{measure_dist_vec}
\alias{measure_dist_vec}
\title{measure distance using haversine}
\usage{
measure_dist_vec(...)
}
\arguments{
\item{...}{argument to pass to geodist::geodist_vec()}
}
\value{
a numeric vector of distances
}
\description{
Wrapper around geodist::geodist_vec that forces haversine method and paired = TRUE
Result is also returned in km instead of m
}
\keyword{internal}
