% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_num_outliers.R
\name{outliers_std_detect}
\alias{outliers_std_detect}
\title{Standard detection of outliers}
\usage{
outliers_std_detect(.data, variable, col_name = raison, filter = TRUE, k = 1.5)
}
\arguments{
\item{.data}{a tibble. if grouped, the analysis will be done per group}

\item{variable}{name of the variable we want to analyse}

\item{col_name}{name of the new column that will be created.}

\item{filter}{a boolean. If TRUE a filter will be applied to keep only outliers.
If FALSE, the number of rows in the output will be the same as .data.}

\item{k}{coefficient applied to the IQR outlier detection}
}
\value{
.data tibble with an additional column names after col_name that conains a string with the type of outlier detection. For non outlier, NA is used
}
\description{
Internal function that detects first centile, last centile and IQR per group
}
\keyword{internal}
