% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_evavelo.R
\name{read_calendrier}
\alias{read_calendrier}
\title{Read and clean "calendrier" information}
\usage{
read_calendrier(file, sheet = "calendrier_sites")
}
\arguments{
\item{file}{xlsx file, Workbook object or URL to xlsx file}

\item{sheet}{Name of the worksheet caontaining "calendrier" information.}
}
\value{
a data.frame
}
\description{
Read a specific sheet of an xlsx object and perform some basic cleaning.
}
