% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_evavelo.R
\name{read_compt_auto}
\alias{read_compt_auto}
\title{Read and clean "comptages_automatiques" information}
\usage{
read_compt_auto(file, sheet = "comptages_automatiques")
}
\arguments{
\item{file}{xlsx file, Workbook object or URL to xlsx file}

\item{sheet}{Name of the worksheet containing "comptages_automatiques" information.}
}
\value{
a data.frame
}
\description{
Read a specific sheet of an xlsx object and perform some basic cleaning.
Compute predictors that are later user for clustering classification
}
