% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_evavelo.R
\name{read_comptage}
\alias{read_comptage}
\title{Read and clean "comptage" information}
\usage{
read_comptage(file, init = FALSE)
}
\arguments{
\item{file}{xlsx file, Workbook object or URL to xlsx file}

\item{init}{boolean to import the initial sheet instead of the _post_traitement}
}
\value{
a data.frame
}
\description{
Read a specific sheet of an xlsx object and perform some basic cleaning.
}
\keyword{internal}
