% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_evavelo.R
\name{read_enquete}
\alias{read_enquete}
\title{Read and clean "enquete" information}
\usage{
read_enquete(file, sheet = "enquetes_post_traitement")
}
\arguments{
\item{file}{xlsx file, Workbook object or URL to xlsx file}

\item{sheet}{Name of the worksheet caontaining "enquete" information.}
}
\value{
a data.frame
}
\description{
Read a specific sheet of an xlsx object and perform some basic cleaning.
}
