% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_evavelo.R
\name{read_evavelo}
\alias{read_evavelo}
\title{Read and clean eva-velo xlsx file}
\usage{
read_evavelo(file)
}
\arguments{
\item{file}{xlsx file, Workbook object or URL to xlsx file}
}
\value{
a list of data.frames
}
\description{
Read specific sheets from an xlsx object and perform some basic cleaning.
}
