% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distance.R
\name{real_distance}
\alias{real_distance}
\title{Compute real distance based on Indiggo formula}
\usage{
real_distance(dist)
}
\arguments{
\item{dist}{a vector of distances to be updated}
}
\value{
a vector the same size as dist with computed value
}
\description{
Compute real distance based on Indiggo formula
}
\keyword{internal}
