% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_french_cities}
\alias{rename_french_cities}
\title{Rename french cities names compatible to local data.base}
\usage{
rename_french_cities(cities)
}
\arguments{
\item{cities}{a string vector}
}
\value{
a string vector the same length as cities.
}
\description{
Convert a string vector and change it to capital without accent. Removes `-` or `'` and transform SAINT to ST.
}
\keyword{internal}
