% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum_prod}
\alias{sum_prod}
\title{Multiply a list of vectors element-wise and and sum the result}
\usage{
sum_prod(..., na.rm = TRUE)
}
\arguments{
\item{...}{a list of vectors}

\item{na.rm}{Remove NA during the summation. Default to TRUE}
}
\value{
a single value
}
\description{
Useful function inside a summarize() to multiply columns together especially if some are booleans.
This way we can summarize in a single command with different "pseudo-filters"
}
\keyword{internal}
