% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distance.R
\name{unify_regions}
\alias{unify_regions}
\title{Combined studied regions together}
\usage{
unify_regions(evadata)
}
\arguments{
\item{evadata}{a geocoded eva_data object that contains tables_communes}
}
\value{
an sf object
}
\description{
Based on "tables_communes", this function lists all regions on the itinerary and create a new sf object with the union of these regions
}
\keyword{internal}
