% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_wb.R
\name{update_wb}
\alias{update_wb}
\title{Update a workbook with data from R object}
\usage{
update_wb(wb, data)
}
\arguments{
\item{wb}{a workbook object created with openxlsx::loadWorkbook()}

\item{data}{a named list of data.frames
Each name correspond to a worksheet name to update
Each data.frame column is corresponding to a column name in wb}
}
\value{
invisible(0)
}
\description{
This function doesn't return an value but modifies object `wb`
}
