% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_wb.R
\name{update_ws}
\alias{update_ws}
\title{Update a specific worksheet from a workbook.}
\usage{
update_ws(wb, data, sheet)
}
\arguments{
\item{wb}{a workbook object created with openxlsx::loadWorkbook()}

\item{data}{a named list of data.frames}

\item{sheet}{name of the sheet that needs to be updated}
}
\value{
invisible(0)
}
\description{
Update a specific worksheet from a workbook.
}
\keyword{internal}
