% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_min}
\alias{which_min}
\title{An improved version of base R which.min that returns NA is all are Nas (instead of empty integer)}
\usage{
which_min(x)
}
\arguments{
\item{x}{numeric vector whose min is searched for}
}
\value{
an integer vector with position of minimum value
}
\description{
An improved version of base R which.min that returns NA is all are Nas (instead of empty integer)
}
\keyword{internal}
