% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distance.R
\name{calc_distance}
\alias{calc_distance}
\title{Computes distance (3.1.19-3.1.21)}
\usage{
calc_distance(eva_data, max_dist = 30)
}
\arguments{
\item{eva_data}{an evadata object obtained using evavelo::read_evavelo()}

\item{max_dist}{distances greater than this value in kilometers will be discarded when looking for iti_dep_iti and iti_arr_iti and result will have NAs}
}
\value{
a list of data.frames with modified columns from the original xlsx file
}
\description{
This function does the following actions
* Assign COG codes to ville_heb_cog_lau and ville_res_cog_lau
* Fill  iti_dep_iti_valide and iti_arr_iti_valide with the closest city on itinary
* Compute distances distance_domicile_enq distance_dom_enq_reelle distance_heb_enq and distance_heb_enq_reelle
}
\keyword{internal}
