% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_evavelo.R
\name{check_evavelo}
\alias{check_evavelo}
\title{Check input file correcteness (Eva-velo)}
\usage{
check_evavelo(comptage, enquete)
}
\arguments{
\item{comptage}{a data.frame}

\item{enquete}{a data.frame}
}
\value{
a list of two values: error (a boolean) and log (a string)
}
\description{
Check input file correcteness (Eva-velo)
}
