% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_evavelo.R
\name{find_wrong_cities}
\alias{find_wrong_cities}
\title{Find cities that couldn't be geocoded by geocode_cities() or that have been interpreted due to a mispelling}
\usage{
find_wrong_cities(
  var_name,
  enquete,
  fields_to_keep = "id_quest",
  include_interpretation = TRUE
)
}
\arguments{
\item{var_name}{string corresponding to the city name variable to check}

\item{enquete}{data.frame containing the var_name variable and an associated *_cog}

\item{fields_to_keep}{fields that will be exported along with var_name}

\item{include_interpretation}{Should we include geocoding interpretation? if FALSE only errors will be output.}
}
\value{
a data.frame with cities that couldn't be geocoded
}
\description{
Find cities that couldn't be geocoded by geocode_cities() or that have been interpreted due to a mispelling
}
\keyword{internal}
