% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_cities.R
\name{geocode_cities}
\alias{geocode_cities}
\title{geocode a data.frame based on city names}
\usage{
geocode_cities(.data, city_col)
}
\arguments{
\item{.data}{a dataframe that needs to be updated}

\item{city_col}{column name containing the city name in .data
This value will also be used to name the new columns.}
}
\value{
the input data.frame with 3 new columns with a name based on city_col (_lat, _long, _cog)
}
\description{
This function will try to find the best match in case of mispelling.
A warning message will be generated indicating the city that has been chosen in case of doubt.
}
