% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_cities.R
\name{geocode_city}
\alias{geocode_city}
\title{Geocode one city name}
\usage{
geocode_city(city_name, score_limit = 0.4)
}
\arguments{
\item{city_name}{a vector of city names}

\item{score_limit}{results below this score will be ignored and replaced by NAs}
}
\value{
a one line data.frame
}
\description{
wrapper around banR::geocode() but that limits to one result of type "municipality"
}
