% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_categ.R
\name{is_iti_coherent}
\alias{is_iti_coherent}
\title{Helper function to control "coherence" of specific "itinerance" answers:}
\usage{
is_iti_coherent(
  dms,
  iti_km_voyage,
  iti_experience,
  iti_depart_itineraire,
  iti_arrivee_itineraire,
  ...
)
}
\arguments{
\item{dms}{numeric vector: 'duree moyenne de sejour'}

\item{iti_km_voyage}{numeric vector}

\item{iti_experience}{character vector}

\item{iti_depart_itineraire}{character vector}

\item{iti_arrivee_itineraire}{character vector}

\item{...}{other answers to iti_* questions}
}
\value{
a boolean vector indicating if answer is coherent
}
\description{
https://github.com/JMPivette/evavelo/discussions/7
}
