% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_evavelo.R
\name{process_evavelo}
\alias{process_evavelo}
\title{Read and process a eva-velo file}
\usage{
process_evavelo(file)
}
\arguments{
\item{file}{An xlsx file, Workbook object or URL to xlsx file}
}
\value{
a list of corrected categories
}
\description{
Read and process a eva-velo file
}
