% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_evavelo.R
\name{read_comptage}
\alias{read_comptage}
\title{Read and clean "comptage" information}
\usage{
read_comptage(file, sheet = "comptages_man_post_traitements")
}
\arguments{
\item{file}{xlsx file, Workbook object or URL to xlsx file}

\item{sheet}{Name of the worksheet caontaining "comptage" information.}
}
\value{
a data.frame
}
\description{
Read a specific sheet of an xlsx object and perform some basic cleaning.
}
